Fixup ping source dataMaxSize configmap
- Configmap should be basic key/value pair, not a string for a yaml struct
- Provide example dataMaxSize configmap config
- Correct example-checksum annotation
- No mashalling from yaml needed, strconv directly to int64
- Add unit testing for ConfigMap reading
- Remove erroneous broker parsing comments/function from copied implementation
- Rename configmap to match naming conventions
- update bsize len assignment to match int64 size

+++ pkg/apis/sources/v1alpha2/ping_validation.go
@@ -43,7 +43,7 @@ func (cs *PingSourceSpec) Validate(ctx context.Context) *apis.FieldError {
 	pingConfig := config.FromContextOrDefaults(ctx)
 	pingDefaults := pingConfig.PingDefaults.GetPingConfig()
 
-	if bsize := len(cs.JsonData); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
+	if bsize := int64(len(cs.JsonData)); pingDefaults.DataMaxSize > -1 && bsize > pingDefaults.DataMaxSize {
 		fe := apis.ErrInvalidValue(fmt.Sprintf("the jsonData length of %d bytes exceeds limit set at %d.", bsize, pingDefaults.DataMaxSize), "jsonData")
 		errs = errs.Also(fe)
 	}
